backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
langchain
langchain-openai
python-dotenv